* BNF grammar for wrightscript defined * Function syntax proposal and example added
The grammar was derived from some scripts of 'Turnabout Substitution'
and the API doc. Some rules had to be dropped for easier parsing, e.g.

sfx ~This is a long songtitle.ogg

because it required special rules to see the following argument as one
big string. With the new grammar, most (if not all) built-in functions
and control structures follow the same syntax and can be implemented
in wrightscript itself or as simple identifier => python function mappings.
The goal is to make libengine.py much shorter and move the parsing
to another module.

The new function syntax allows for function definitions inside wrightscript.
They are defined like macros, yet they may return a value and the arguments
are explicitly named.