Added entity syntax.
An entity is similar to a struct in C/C++. They define a list of fields
that have to be present when creating an instance of this entity. It
is possible to state default values and also possible to add fields on
instance creation or later. Examples:

<code>
entity Globals
  preload = false
  music_fade = 100
  # ...
endentity

#...

entity Character
  name
  blipsound
endentity

ema := Character {
           name := "Ema",
           blipsound := Femblib, # assuming that Femblib is itself a defined "Sound" entity
           mad := Portrait { # nested entities
                    file := "/ema_mad.png"
                    #...
           }
       }

portrait ema.mad
lipsync true
speaker ema # may set name tag + blipsound
"Hello World!"
</code>